Add files via upload
Zip contains an .exe file which runs a program to add a list to a specified SharePoint site collection. You choose the number of items. The list is from "Custom List" template with added columns: Body, SomeText, SomeNumber, EmailAddress.
SPO has an unchangeable view limit for lists and it is 5000 elements. With this list you can go way above the SharePoint limit of 5000 items (tested with 12 000), so if you decide to create a list with 40 000 elements, expect some issues there. For your convenience, I created a default view with 70 items limit. You can easily switch between the views in the GUI. 

Some of the fields intentionally have the same values to let you test filtering in your SharePoint Online tenant.

The program is meant mainly for  testing purposes, e.g.

- creating a test environment for your company

- recreating big data and a lot of items without resorting to Excel and QuickEdit

- troubleshooting issues which occur when your users added more than 5000 items and you don't want or can't test on the real data

 

As it runs, it will inform you on the percentage of its progress.

 

 

Everything was written in C# and sorry for this, but I only tested on environments with SharePoint SDK installed, so if SDK turns out to be an unintentional prerequisite, I do apologize.

 

Below you can find part of the code:

C#
string strona = site ; 
            var context = new ClientContext(strona); 
            context.Credentials = new SharePointOnlineCredentials(credUsername, password); 
            context.Load(context.Web.Webs, wc => wc.Include(w => w.Url, w => w.Lists)); 
            context.ExecuteQuery(); 
            context.Web.Lists.Add(CreateList("Custom List with "+itemsno.ToString()+ " elements", 100)); 
            context.ExecuteQuery(); 
            context.Load(context.Web.Webs, wc => wc.Include(w => w.Url, w => w.Lists)); 
            context.ExecuteQuery(); 
            Microsoft.SharePoint.Client.List generici = context.Web.Lists.GetByTitle("Custom List with " + itemsno.ToString() + " elements"); 
            generici.Fields.AddFieldAsXml("<Field Type='Text' DisplayName='Body'/>", true, AddFieldOptions.AddToAllContentTypes); 
            generici.Fields.AddFieldAsXml("<Field Type='Text' DisplayName='SomeText'/>", true, AddFieldOptions.AddToAllContentTypes); 
            generici.Fields.AddFieldAsXml("<Field Type='Number' DisplayName='SomeNumber'/>", true, AddFieldOptions.AddToAllContentTypes); 
            generici.Fields.AddFieldAsXml("<Field Type='Text' DisplayName='EmailAddress'/>", true, AddFieldOptions.AddToAllContentTypes); 
            generici.Update(); 
            //context.ExecuteQuery(); 
 
 
            ViewCreationInformation vci = new ViewCreationInformation(); 
            vci.RowLimit = 70; 
            vci.Title = "Modified View with 70 items limit"; 
            string[] vf = { "SomeText", "EmailAddress", "SomeNumber", "Body" }; 
            vci.ViewFields = vf; 
            vci.SetAsDefaultView = true; 
            generici.Views.Add(vci); 
            context.ExecuteQuery(); 
            double loopno = (double)itemsno / 4; 
            loopno = Math.Ceiling(loopno); 
            Console.WriteLine("\n\nThe list has been created. Moving on to populate it..."); 
            Random rand = new Random(); 
            for (int i = 0; i < loopno; i++) 
            {